Revert "feat: manage groups dialog improvement"
