+++ src/app/components/manage-groups-dialog.component.ts
@@ -6,14 +6,13 @@ import { MatButtonModule } from '@angular/material/button';
 import { MAT_DIALOG_DATA, MatDialog, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatTooltipModule } from '@angular/material/tooltip';
 import { ManageGroupsDialogData, ManageGroupsDialogResult, TasksStatusesGroup } from '@app/dashboard/types';
 import { TasksStatusesService } from '@app/tasks/services/tasks-statuses.service';
 import { ActionsToolbarGroupComponent } from '@components/actions-toolbar-group.component';
 import { ActionsToolbarComponent } from '@components/actions-toolbar.component';
 import { IconsService } from '@services/icons.service';
-import { AddStatusesGroupDialogComponent } from './add-statuses-group-dialog.component';
-import { EditStatusesGroupDialogComponent } from './edit-status-group-dialog.component';
+import { AddStatusesGroupDialogComponent } from '../dashboard/components/add-statuses-group-dialog.component';
+import { EditStatusesGroupDialogComponent } from '../dashboard/components/edit-status-group-dialog.component';
 
 @Component({
   templateUrl: './manage-groups-dialog.component.html',
@@ -25,6 +24,8 @@ ul {
 }
 
 .groups {
+  margin-top: 1rem;
+
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   grid-gap: 2rem;
@@ -42,25 +43,6 @@ ul {
 
   padding: 0.5rem;
   border-radius: 0.5rem;
-}
-
-.addGroup {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.addGroup button {
-  display: flex;
-  flex-direction: column;
-  min-height: 5rem;
-  width: 80%;
-}
-
-.addGroup button mat-icon {
-  transform: scale(1.5);
-  margin: 0 !important;
-  margin-top: 0.5rem !important;
 }
 
 .groups ul li mat-icon {
@@ -126,8 +108,7 @@ ul {
     MatDialogModule,
     MatToolbarModule,
     MatIconModule,
-    DragDropModule,
-    MatTooltipModule,
+    DragDropModule
   ]
 })
 export class ManageGroupsDialogComponent implements OnInit {
